requirements.txt
streamlit>=1.28.0
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
seaborn>=0.12.0
openpyxl>=3.1.0
pyarrow>=12.0.0
xlsxwriter
google-generativeai>=0.3.0
python-pptx>=0.6.21